requirements.txt
streamlit
requests
pillow
fpdf
